feat: introduce `Runtime`-wide instance config
